requirements.txt
fastapi
pydantic
uvicorn
redis
requests
python-dotenv
python-multipart
python-jose
passlib